+++ src/commands/key.rs
@@ -8,8 +8,13 @@ use clap::Args;
 
 #[derive(Debug, Default, Args)]
 pub struct KeyCmd {
+    /// Show the current API key.
+    #[arg(long)]
+    show: bool,
+
     /// The API key to set.
-    api_key: String,
+    #[arg(required_unless_present = "show")]
+    api_key: Option<String>,
 
     /// Do not check the authenticity of the key.
     #[arg(short, long)]
@@ -18,25 +23,31 @@ pub struct KeyCmd {
 
 impl Runnable for KeyCmd {
     fn run(self, _flags: &Flags, style: &Style) -> Result<()> {
-        if !self.no_check {
-            println!("Testing key... (use --no-check to skip)");
-            let client = ApiClientManager::new(&self.api_key);
-
-            let id = YoutubeId {
-                id: "dQw4w9WgXcQ".to_string(),
-                is_playlist: false,
-            };
-            if let Err(_) = client.fetch_duration_from_id(&id, 1) {
-                bail!("{}Invalid API key passed!{}", style.red, style.reset)
-            }
-        }
-
         let mut config = Config::load().map_err(|e| anyhow::anyhow!(e.to_string()))?;
 
-        config
-            .update_write_key(self.api_key)
-            .map_err(|e| anyhow::anyhow!(e.to_string()))?;
-        println!("Key added successfully.");
+        if self.show {
+            return Ok(println!("{}", config.api_key().unwrap_or("No key set.")));
+        }
+
+        if let Some(k) = self.api_key {
+            if !self.no_check {
+                println!("Testing key... (use --no-check to skip)");
+                let client = ApiClientManager::new(&k);
+
+                let id = YoutubeId {
+                    id: "dQw4w9WgXcQ".to_string(),
+                    is_playlist: false,
+                };
+                if let Err(_) = client.fetch_duration_from_id(&id, 1) {
+                    bail!("{}Invalid API key passed!{}", style.red, style.reset)
+                }
+            }
+
+            config
+                .update_write_key(k)
+                .map_err(|e| anyhow::anyhow!(e.to_string()))?;
+            println!("Key added successfully.");
+        }
 
         Ok(())
     }